refactor(settings): split oversized settings screens
- SettingsScreen.kt:        1267 -> 905 (5 sibling files - primary fun stays large)
- SettingsSubPages.kt:      1258 -> 396 (9 sibling files, one per sub-screen)
- APISettingsScreen.kt:     1076 -> 369 (8 per-provider section files)
- MediaSettingsScreen.kt:   1067 -> 566 (5 sibling files)
- ParentalControlsScreen.kt: 832 -> 372 (6 sibling files)

Note: SettingsScreen.kt's primary @Composable is structurally one large
block; further decomposition would require pulling sections out of the
single Column, which is left as follow-up work.
