+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/settings/MediaSettingsScreen.kt
@@ -435,426 +435,12 @@ fun MediaSettingsScreen(
 // API PROVIDERS SETTINGS
 // =============================================================================
 
-@Composable
-fun ApiProvidersSettingsScreen(
-    state: ApiProvidersState,
-    onProviderClick: (ApiProvider) -> Unit,
-    onBackClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(MediaColors.Background)
-    ) {
-        SettingsTopBar(
-            title = "API Providers",
-            onBackClick = onBackClick
-        )
-        
-        LazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            contentPadding = PaddingValues(MediaSpacing.ScreenHorizontal),
-            verticalArrangement = Arrangement.spacedBy(MediaSpacing.MD)
-        ) {
-            // Group providers by category
-            val providersByCategory = state.allProviders.groupBy { provider ->
-                when {
-                    provider.supportedCapabilities.any { 
-                        it in listOf(ApiCapability.BOOK_METADATA, ApiCapability.AUDIOBOOK_METADATA) 
-                    } -> "Books & Audiobooks"
-                    provider.supportedCapabilities.any { 
-                        it in listOf(ApiCapability.MOVIE_METADATA, ApiCapability.TV_METADATA) 
-                    } -> "Movies & TV"
-                    provider.supportedCapabilities.any { 
-                        it in listOf(ApiCapability.MUSIC_METADATA, ApiCapability.SCROBBLING) 
-                    } -> "Music"
-                    provider.supportedCapabilities.any { 
-                        it == ApiCapability.PODCAST_METADATA 
-                    } -> "Podcasts"
-                    provider.supportedCapabilities.any { 
-                        it in listOf(ApiCapability.LIBRARY_SYNC, ApiCapability.VIDEO_STREAM) 
-                    } -> "Media Servers"
-                    provider.supportedCapabilities.any { 
-                        it in listOf(ApiCapability.TTS, ApiCapability.TRANSLATION) 
-                    } -> "AI & Cloud Services"
-                    else -> "Other"
-                }
-            }
-            
-            providersByCategory.forEach { (category, providers) ->
-                item {
-                    SettingsSectionHeader(title = category)
-                }
-                
-                item {
-                    SettingsCard {
-                        providers.forEachIndexed { index, provider ->
-                            val isConfigured = provider.id in state.configuredProviderIds
-                            
-                            ApiProviderItem(
-                                provider = provider,
-                                isConfigured = isConfigured,
-                                onClick = { onProviderClick(provider) }
-                            )
-                            
-                            if (index < providers.size - 1) {
-                                SettingsDivider()
-                            }
-                        }
-                    }
-                }
-                
-                item {
-                    Spacer(modifier = Modifier.height(MediaSpacing.MD))
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun ApiProviderItem(
-    provider: ApiProvider,
-    isConfigured: Boolean,
-    onClick: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .padding(MediaSpacing.MD),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Provider icon placeholder
-        Surface(
-            shape = RoundedCornerShape(MediaCorners.SM),
-            color = if (isConfigured) MediaColors.Success.copy(alpha = 0.15f) 
-                   else MediaColors.BackgroundSurface,
-            modifier = Modifier.size(48.dp)
-        ) {
-            Icon(
-                imageVector = if (isConfigured) Icons.Filled.CheckCircle else Icons.Outlined.Api,
-                contentDescription = "Media image",
-                modifier = Modifier.padding(MediaSpacing.SM),
-                tint = if (isConfigured) MediaColors.Success else MediaColors.TextSecondary
-            )
-        }
-        
-        Spacer(modifier = Modifier.width(MediaSpacing.MD))
-        
-        Column(modifier = Modifier.weight(1f)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Text(
-                    text = provider.name,
-                    style = MediaTypography.BodyLarge,
-                    color = MediaColors.TextPrimary,
-                    fontWeight = FontWeight.Medium
-                )
-                
-                if (provider.requiresApiKey || provider.requiresOAuth) {
-                    Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                    Surface(
-                        shape = RoundedCornerShape(MediaCorners.XS),
-                        color = MediaColors.Warning.copy(alpha = 0.15f)
-                    ) {
-                        Text(
-                            text = if (provider.requiresOAuth) "OAuth" else "API Key",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.Warning,
-                            modifier = Modifier.padding(horizontal = MediaSpacing.XS, vertical = 2.dp)
-                        )
-                    }
-                }
-                
-                if (!provider.isFree) {
-                    Spacer(modifier = Modifier.width(MediaSpacing.XS))
-                    Surface(
-                        shape = RoundedCornerShape(MediaCorners.XS),
-                        color = MediaColors.AccentPrimary.copy(alpha = 0.15f)
-                    ) {
-                        Text(
-                            text = "Premium",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.AccentPrimary,
-                            modifier = Modifier.padding(horizontal = MediaSpacing.XS, vertical = 2.dp)
-                        )
-                    }
-                }
-            }
-            
-            Text(
-                text = provider.description,
-                style = MediaTypography.BodySmall,
-                color = MediaColors.TextSecondary
-            )
-            
-            // Show capabilities count
-            Text(
-                text = "${provider.supportedCapabilities.size} capabilities",
-                style = MediaTypography.LabelSmall,
-                color = MediaColors.TextTertiary
-            )
-        }
-        
-        Icon(
-            imageVector = Icons.Default.ChevronRight,
-            contentDescription = "Media image",
-            tint = MediaColors.TextTertiary
-        )
-    }
-}
 
 // =============================================================================
 // PROVIDER CONFIGURATION SCREEN
 // =============================================================================
 
-@Composable
-fun ProviderConfigurationScreen(
-    provider: ApiProvider,
-    configuration: ApiConfiguration?,
-    onSave: (ApiConfiguration) -> Unit,
-    onDelete: () -> Unit,
-    onBackClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var apiKey by remember { mutableStateOf(configuration?.apiKey ?: "") }
-    var serverUrl by remember { mutableStateOf(configuration?.serverUrl ?: "") }
-    var enabledCapabilities by remember { 
-        mutableStateOf(configuration?.enabledCapabilities ?: provider.supportedCapabilities) 
-    }
-    var isEnabled by remember { mutableStateOf(configuration?.isEnabled ?: true) }
-    var showApiKey by remember { mutableStateOf(false) }
-    
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(MediaColors.Background)
-    ) {
-        SettingsTopBar(
-            title = provider.name,
-            onBackClick = onBackClick,
-            actions = {
-                if (configuration != null) {
-                    IconButton(onClick = onDelete) {
-                        Icon(
-                            imageVector = Icons.Default.Delete,
-                            contentDescription = "Delete",
-                            tint = MediaColors.Error
-                        )
-                    }
-                }
-            }
-        )
-        
-        LazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            contentPadding = PaddingValues(MediaSpacing.ScreenHorizontal),
-            verticalArrangement = Arrangement.spacedBy(MediaSpacing.MD)
-        ) {
-            // Provider info
-            item {
-                Surface(
-                    shape = RoundedCornerShape(MediaCorners.Card),
-                    color = MediaColors.BackgroundElevated
-                ) {
-                    Column(
-                        modifier = Modifier.padding(MediaSpacing.MD)
-                    ) {
-                        Text(
-                            text = provider.description,
-                            style = MediaTypography.BodyMedium,
-                            color = MediaColors.TextSecondary
-                        )
-                        
-                        Spacer(modifier = Modifier.height(MediaSpacing.SM))
-                        
-                        Text(
-                            text = provider.website,
-                            style = MediaTypography.BodySmall,
-                            color = MediaColors.AccentPrimary
-                        )
-                    }
-                }
-            }
-            
-            // Enable/Disable toggle
-            item {
-                SettingsCard {
-                    SettingsToggleItem(
-                        title = "Enable ${provider.name}",
-                        subtitle = "Use this provider for configured capabilities",
-                        isChecked = isEnabled,
-                        onCheckedChange = { isEnabled = it }
-                    )
-                }
-            }
-            
-            // Credentials section
-            if (provider.requiresApiKey || provider.requiresOAuth || provider.id in listOf("plex", "jellyfin", "emby", "calibre")) {
-                item {
-                    Spacer(modifier = Modifier.height(MediaSpacing.MD))
-                    SettingsSectionHeader(title = "Credentials")
-                }
-                
-                item {
-                    SettingsCard {
-                        if (provider.requiresApiKey) {
-                            Column(modifier = Modifier.padding(MediaSpacing.MD)) {
-                                Text(
-                                    text = "API Key",
-                                    style = MediaTypography.LabelMedium,
-                                    color = MediaColors.TextSecondary
-                                )
-                                
-                                Spacer(modifier = Modifier.height(MediaSpacing.SM))
-                                
-                                OutlinedTextField(
-                                    value = apiKey,
-                                    onValueChange = { apiKey = it },
-                                    modifier = Modifier.fillMaxWidth(),
-                                    placeholder = { Text("Enter your API key") },
-                                    visualTransformation = if (showApiKey) 
-                                        VisualTransformation.None 
-                                    else 
-                                        PasswordVisualTransformation(),
-                                    trailingIcon = {
-                                        IconButton(onClick = { showApiKey = !showApiKey }) {
-                                            Icon(
-                                                imageVector = if (showApiKey) 
-                                                    Icons.Default.VisibilityOff 
-                                                else 
-                                                    Icons.Default.Visibility,
-                                                contentDescription = "Media image"
-                                            )
-                                        }
-                                    },
-                                    colors = OutlinedTextFieldDefaults.colors(
-                                        focusedBorderColor = MediaColors.AccentPrimary,
-                                        unfocusedBorderColor = MediaColors.Border
-                                    ),
-                                    shape = RoundedCornerShape(MediaCorners.SM)
-                                )
-                            }
-                        }
-                        
-                        if (provider.id in listOf("plex", "jellyfin", "emby", "calibre")) {
-                            if (provider.requiresApiKey) {
-                                SettingsDivider()
-                            }
-                            
-                            Column(modifier = Modifier.padding(MediaSpacing.MD)) {
-                                Text(
-                                    text = "Server URL",
-                                    style = MediaTypography.LabelMedium,
-                                    color = MediaColors.TextSecondary
-                                )
-                                
-                                Spacer(modifier = Modifier.height(MediaSpacing.SM))
-                                
-                                OutlinedTextField(
-                                    value = serverUrl,
-                                    onValueChange = { serverUrl = it },
-                                    modifier = Modifier.fillMaxWidth(),
-                                    placeholder = { Text("https://your-server.com") },
-                                    colors = OutlinedTextFieldDefaults.colors(
-                                        focusedBorderColor = MediaColors.AccentPrimary,
-                                        unfocusedBorderColor = MediaColors.Border
-                                    ),
-                                    shape = RoundedCornerShape(MediaCorners.SM)
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-            
-            // Capabilities section
-            item {
-                Spacer(modifier = Modifier.height(MediaSpacing.MD))
-                SettingsSectionHeader(title = "Capabilities")
-                
-                Text(
-                    text = "Select which features to use from ${provider.name}",
-                    style = MediaTypography.BodySmall,
-                    color = MediaColors.TextTertiary,
-                    modifier = Modifier.padding(top = MediaSpacing.XS)
-                )
-            }
-            
-            // Group capabilities by category
-            val capabilitiesByCategory = provider.supportedCapabilities.groupBy { it.category }
-            
-            capabilitiesByCategory.forEach { (category, capabilities) ->
-                item {
-                    Text(
-                        text = category.displayName,
-                        style = MediaTypography.LabelMedium,
-                        color = MediaColors.TextSecondary,
-                        modifier = Modifier.padding(top = MediaSpacing.MD, bottom = MediaSpacing.XS)
-                    )
-                }
-                
-                item {
-                    SettingsCard {
-                        capabilities.forEachIndexed { index, capability ->
-                            SettingsToggleItem(
-                                title = capability.displayName,
-                                subtitle = capability.description,
-                                isChecked = capability in enabledCapabilities,
-                                onCheckedChange = { checked ->
-                                    enabledCapabilities = if (checked) {
-                                        enabledCapabilities + capability
-                                    } else {
-                                        enabledCapabilities - capability
-                                    }
-                                }
-                            )
-                            
-                            if (index < capabilities.size - 1) {
-                                SettingsDivider()
-                            }
-                        }
-                    }
-                }
-            }
-            
-            // Save button
-            item {
-                Spacer(modifier = Modifier.height(MediaSpacing.LG))
-                
-                Button(
-                    onClick = {
-                        onSave(
-                            ApiConfiguration(
-                                providerId = provider.id,
-                                apiKey = apiKey.takeIf { it.isNotBlank() },
-                                serverUrl = serverUrl.takeIf { it.isNotBlank() },
-                                enabledCapabilities = enabledCapabilities,
-                                isEnabled = isEnabled
-                            )
-                        )
-                    },
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = MediaColors.AccentPrimary
-                    ),
-                    shape = RoundedCornerShape(MediaCorners.Button)
-                ) {
-                    Text(
-                        text = "Save Configuration",
-                        style = MediaTypography.LabelLarge,
-                        modifier = Modifier.padding(vertical = MediaSpacing.SM)
-                    )
-                }
-            }
-            
-            item {
-                Spacer(modifier = Modifier.height(MediaSpacing.Huge))
-            }
-        }
-    }
-}
 
 // =============================================================================
 // SETTINGS COMPONENTS
@@ -921,94 +507,7 @@ private fun SettingsCard(
     }
 }
 
-@Composable
-private fun SettingsItem(
-    icon: ImageVector,
-    iconColor: Color,
-    title: String,
-    subtitle: String,
-    onClick: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .padding(MediaSpacing.MD),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Surface(
-            shape = CircleShape,
-            color = iconColor.copy(alpha = 0.15f),
-            modifier = Modifier.size(40.dp)
-        ) {
-            Icon(
-                imageVector = icon,
-                contentDescription = "Media image",
-                modifier = Modifier.padding(MediaSpacing.SM),
-                tint = iconColor
-            )
-        }
-        
-        Spacer(modifier = Modifier.width(MediaSpacing.MD))
-        
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = title,
-                style = MediaTypography.BodyLarge,
-                color = MediaColors.TextPrimary
-            )
-            Text(
-                text = subtitle,
-                style = MediaTypography.BodySmall,
-                color = MediaColors.TextSecondary
-            )
-        }
-        
-        Icon(
-            imageVector = Icons.Default.ChevronRight,
-            contentDescription = "Media image",
-            tint = MediaColors.TextTertiary
-        )
-    }
-}
 
-@Composable
-private fun SettingsToggleItem(
-    title: String,
-    subtitle: String,
-    isChecked: Boolean,
-    onCheckedChange: (Boolean) -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onCheckedChange(!isChecked) }
-            .padding(MediaSpacing.MD),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = title,
-                style = MediaTypography.BodyLarge,
-                color = MediaColors.TextPrimary
-            )
-            Text(
-                text = subtitle,
-                style = MediaTypography.BodySmall,
-                color = MediaColors.TextSecondary
-            )
-        }
-        
-        Switch(
-            checked = isChecked,
-            onCheckedChange = onCheckedChange,
-            colors = SwitchDefaults.colors(
-                checkedThumbColor = MediaColors.AccentPrimary,
-                checkedTrackColor = MediaColors.AccentPrimary.copy(alpha = 0.5f)
-            )
-        )
-    }
-}
 
 @Composable
 private fun SettingsDivider() {
