+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/settings/AutoScanSettingsScreen.kt
@@ -1,0 +1,179 @@
+package com.universalmedialibrary.ui.settings
+
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
+import android.widget.Toast
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.unit.dp
+import androidx.core.content.FileProvider
+import androidx.core.content.ContextCompat
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.viewModelScope
+import androidx.navigation.NavController
+import com.universalmedialibrary.data.repository.CacheLocation
+import com.universalmedialibrary.data.repository.SettingsRepository
+import com.universalmedialibrary.data.settings.BottomGearPosition
+import com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode
+import com.universalmedialibrary.services.MediaScannerService
+import com.universalmedialibrary.services.debug.DebugBugReportService
+import com.universalmedialibrary.utils.PermissionsHandler
+import com.universalmedialibrary.utils.rememberPermissionsHandler
+import com.universalmedialibrary.jobs.WorkScheduler
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import java.io.File
+import javax.inject.Inject
+import java.util.Locale
+import java.util.concurrent.TimeUnit
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun AutoScanSettingsScreen(
+    onBack: () -> Unit,
+    viewModel: AutoScanSettingsViewModel = hiltViewModel()
+) {
+    val context = LocalContext.current
+    val state by viewModel.state.collectAsState()
+    val scope = rememberCoroutineScope()
+    val permissions = rememberPermissionsHandler()
+
+    fun rescheduleWork() {
+        if (!state.enabled) {
+            WorkScheduler.cancelPeriodicMetadataRefresh(context)
+            return
+        }
+
+        WorkScheduler.schedulePeriodicMetadataRefresh(
+            context = context,
+            intervalHours = state.intervalHours.toLong(),
+            wifiOnly = state.wifiOnly
+        )
+    }
+
+    // Keep WorkManager schedule in sync with stored settings.
+    LaunchedEffect(state.enabled, state.intervalHours, state.wifiOnly) {
+        // Debounce rapid successive changes (toggle + interval edits).
+        delay(500)
+        rescheduleWork()
+    }
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text("Auto-Scan") },
+                navigationIcon = {
+                    IconButton(onClick = onBack) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                },
+                actions = {
+                    IconButton(
+                        onClick = { rescheduleWork() }
+                    ) {
+                        Icon(Icons.Default.Refresh, contentDescription = "Reschedule")
+                    }
+                }
+            )
+        }
+    ) { padding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(padding)
+                .verticalScroll(rememberScrollState())
+                .padding(16.dp),
+            verticalArrangement = Arrangement.spacedBy(16.dp)
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text("Enable periodic scans", fontWeight = FontWeight.Medium)
+                    Text(
+                        "Automatically scan for new media in the background",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+                Switch(checked = state.enabled, onCheckedChange = viewModel::setEnabled)
+            }
+
+            OutlinedTextField(
+                value = state.intervalHours.toString(),
+                onValueChange = { raw ->
+                    raw.toIntOrNull()?.let { hours ->
+                        if (hours >= 1) viewModel.setIntervalHours(hours)
+                    }
+                },
+                label = { Text("Interval (hours)") },
+                supportingText = { Text("Minimum 1 hour") },
+                leadingIcon = { Icon(Icons.Default.Schedule, contentDescription = "Scan interval") },
+                enabled = state.enabled,
+                modifier = Modifier.fillMaxWidth()
+            )
+
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text("Wi‑Fi only", fontWeight = FontWeight.Medium)
+                    Text(
+                        "Avoid scanning on metered networks",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+                Switch(checked = state.wifiOnly, onCheckedChange = viewModel::setWifiOnly, enabled = state.enabled)
+            }
+
+            Button(
+                onClick = {
+                    // Validate storage permissions before starting scan to avoid SecurityException.
+                    if (!PermissionsHandler.hasStoragePermissions(context)) {
+                        permissions.requestPermissions()
+                        Toast.makeText(context, "Storage permissions required to scan.", Toast.LENGTH_SHORT).show()
+                        return@Button
+                    }
+
+                    val intent = Intent(context, MediaScannerService::class.java).apply {
+                        action = MediaScannerService.ACTION_SCAN_ALL
+                    }
+                    ContextCompat.startForegroundService(context, intent)
+                },
+                modifier = Modifier.fillMaxWidth()
+            ) {
+                Icon(Icons.Default.Search, contentDescription = "Scan")
+                Spacer(modifier = Modifier.width(8.dp))
+                Text("Scan now")
+            }
+
+            Text(
+                text = "Auto-scan uses the scheduler layer and runs a foreground scan job when scheduled.",
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+        }
+    }
+}
