+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/settings/CastingSettingsScreen.kt
@@ -1,0 +1,112 @@
+package com.universalmedialibrary.ui.settings
+
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
+import android.widget.Toast
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.unit.dp
+import androidx.core.content.FileProvider
+import androidx.core.content.ContextCompat
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.viewModelScope
+import androidx.navigation.NavController
+import com.universalmedialibrary.data.repository.CacheLocation
+import com.universalmedialibrary.data.repository.SettingsRepository
+import com.universalmedialibrary.data.settings.BottomGearPosition
+import com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode
+import com.universalmedialibrary.services.MediaScannerService
+import com.universalmedialibrary.services.debug.DebugBugReportService
+import com.universalmedialibrary.utils.PermissionsHandler
+import com.universalmedialibrary.utils.rememberPermissionsHandler
+import com.universalmedialibrary.jobs.WorkScheduler
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import java.io.File
+import javax.inject.Inject
+import java.util.Locale
+import java.util.concurrent.TimeUnit
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CastingSettingsScreen(
+    onBack: () -> Unit,
+    viewModel: CastingSettingsViewModel = hiltViewModel()
+) {
+    val state by viewModel.state.collectAsState()
+    val context = LocalContext.current
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text("Casting") },
+                navigationIcon = {
+                    IconButton(onClick = onBack) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                }
+            )
+        }
+    ) { padding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(padding)
+                .verticalScroll(rememberScrollState())
+                .padding(16.dp),
+            verticalArrangement = Arrangement.spacedBy(16.dp)
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text("Enable casting", fontWeight = FontWeight.Medium)
+                    Text(
+                        "Show cast targets and allow casting audio",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+                Switch(checked = state.enabled, onCheckedChange = viewModel::setEnabled)
+            }
+
+            Button(
+                onClick = {
+                    // Open system Cast settings if present; otherwise open Bluetooth settings as a fallback.
+                    val intent = Intent(android.provider.Settings.ACTION_CAST_SETTINGS)
+                    runCatching { context.startActivity(intent) }.onFailure {
+                        Toast.makeText(
+                            context,
+                            "Cast settings not available on this device",
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    }
+                },
+                modifier = Modifier.fillMaxWidth()
+            ) {
+                Icon(Icons.Default.Settings, contentDescription = "Cast settings")
+                Spacer(modifier = Modifier.width(8.dp))
+                Text("Open system cast settings")
+            }
+        }
+    }
+}
