+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/settings/ParentalControlsScreen.kt
@@ -274,183 +274,7 @@ fun ParentalControlsScreen(
     }
 }
 
-@Composable
-private fun PinSetupCard(
-    hasPinSet: Boolean,
-    onSetPin: () -> Unit,
-    onChangePin: () -> Unit,
-    onClearPin: () -> Unit
-) {
-    Card(modifier = Modifier.fillMaxWidth()) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Icon(
-                    if (hasPinSet) Icons.Default.Lock else Icons.Default.LockOpen,
-                    contentDescription = "Media image",
-                    tint = if (hasPinSet) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.error
-                )
-                Text(
-                    "PIN Protection",
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Medium
-                )
-            }
 
-            Text(
-                if (hasPinSet) "PIN is set and protecting your content" else "Set a PIN to lock adult content and settings",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                if (!hasPinSet) {
-                    Button(
-                        onClick = onSetPin,
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Icon(Icons.Default.Add, contentDescription = "Add", modifier = Modifier.size(18.dp))
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text("Set PIN")
-                    }
-                } else {
-                    OutlinedButton(
-                        onClick = onChangePin,
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Icon(Icons.Default.Edit, contentDescription = "Edit", modifier = Modifier.size(18.dp))
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text("Change")
-                    }
-                    OutlinedButton(
-                        onClick = onClearPin,
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Icon(Icons.Default.Delete, contentDescription = "Delete", modifier = Modifier.size(18.dp))
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text("Clear")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ContentRestrictionsCard(
-    blockMature: Boolean,
-    blockExplicit: Boolean,
-    hideAdultContent: Boolean,
-    requirePinForAdult: Boolean,
-    allowAdultSources: Boolean,
-    canToggleAdultSources: Boolean,
-    adultSourcesDisabledReason: String?,
-    hasPinSet: Boolean,
-    onBlockMatureChange: (Boolean) -> Unit,
-    onBlockExplicitChange: (Boolean) -> Unit,
-    onHideAdultContentChange: (Boolean) -> Unit,
-    onAllowAdultSourcesChange: (Boolean) -> Unit,
-    onRequirePinChange: (Boolean) -> Unit
-) {
-    Card(modifier = Modifier.fillMaxWidth()) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Icon(
-                    Icons.Default.Block,
-                    contentDescription = "Media image",
-                    tint = MaterialTheme.colorScheme.error
-                )
-                Text(
-                    "Content Restrictions",
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Medium
-                )
-            }
-
-            // Block Explicit
-            SwitchSettingItem(
-                title = "Block Explicit Content",
-                description = "Block stories rated Explicit/Adult/NC-17",
-                checked = blockExplicit,
-                onCheckedChange = onBlockExplicitChange,
-                icon = Icons.Default.Warning
-            )
-
-            HorizontalDivider()
-
-            // Block Mature
-            SwitchSettingItem(
-                title = "Block Mature Content",
-                description = "Block stories rated Mature/M/R",
-                checked = blockMature,
-                onCheckedChange = onBlockMatureChange,
-                icon = Icons.Default.WarningAmber
-            )
-
-            HorizontalDivider()
-
-            // Hide Adult Content
-            SwitchSettingItem(
-                title = "Hide Adult Content",
-                description = "Completely hide blocked content from library and searches",
-                checked = hideAdultContent,
-                onCheckedChange = onHideAdultContentChange,
-                icon = Icons.Default.VisibilityOff
-            )
-
-            HorizontalDivider()
-
-            // Adult Content Sources toggle
-            val adultDescription = if (canToggleAdultSources) {
-                "Allow searching and downloading from adult-oriented fiction sites"
-            } else {
-                adultSourcesDisabledReason ?: "Currently disabled by parental control settings"
-            }
-            SwitchSettingItem(
-                title = "Enable Adult Story Sources",
-                description = adultDescription,
-                checked = allowAdultSources && canToggleAdultSources,
-                onCheckedChange = onAllowAdultSourcesChange,
-                icon = Icons.Default.Explicit,
-                enabled = canToggleAdultSources
-            )
-
-            if (!canToggleAdultSources) {
-                Text(
-                    text = adultSourcesDisabledReason ?: "Adult site access is disabled by current settings.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            HorizontalDivider()
-
-            // Require PIN for Adult
-            SwitchSettingItem(
-                title = "Require PIN for Adult Content",
-                description = if (hasPinSet) "PIN required to access mature/explicit content" else "Set a PIN first to enable",
-                checked = requirePinForAdult,
-                onCheckedChange = onRequirePinChange,
-                icon = Icons.Default.Lock,
-                enabled = hasPinSet
-            )
-        }
-    }
-}
 
 @Composable
 private fun AdditionalSettingsCard(
@@ -488,65 +312,6 @@ private fun AdditionalSettingsCard(
     }
 }
 
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun MediaRatingsCard(
-    movieLimit: ParentalControlsSettings.MovieRatingLevel,
-    tvLimit: ParentalControlsSettings.TvRatingLevel,
-    gameLimit: ParentalControlsSettings.GameRatingLevel,
-    bookLimit: ParentalControlsSettings.BookRatingLevel,
-    onMovieLimitChange: (ParentalControlsSettings.MovieRatingLevel) -> Unit,
-    onTvLimitChange: (ParentalControlsSettings.TvRatingLevel) -> Unit,
-    onGameLimitChange: (ParentalControlsSettings.GameRatingLevel) -> Unit,
-    onBookLimitChange: (ParentalControlsSettings.BookRatingLevel) -> Unit
-) {
-    Card(modifier = Modifier.fillMaxWidth()) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(20.dp)
-        ) {
-            RatingLimitSelector<ParentalControlsSettings.MovieRatingLevel>(
-                icon = Icons.Default.Movie,
-                title = "Films & Movies",
-                subtitle = "Set the highest MPAA rating that can be accessed.",
-                options = ParentalControlsSettings.MovieRatingLevel.values(),
-                selected = movieLimit,
-                labelProvider = { it.displayName },
-                onSelected = onMovieLimitChange
-            )
-
-            RatingLimitSelector<ParentalControlsSettings.TvRatingLevel>(
-                icon = Icons.Default.Tv,
-                title = "TV & Streaming",
-                subtitle = "Choose the maximum TV rating (TV-Y through TV-MA).",
-                options = ParentalControlsSettings.TvRatingLevel.values(),
-                selected = tvLimit,
-                labelProvider = { it.displayName },
-                onSelected = onTvLimitChange
-            )
-
-            RatingLimitSelector<ParentalControlsSettings.GameRatingLevel>(
-                icon = Icons.Default.SportsEsports,
-                title = "Games & Interactive Media",
-                subtitle = "Limit ESRB ratings for games and interactive content.",
-                options = ParentalControlsSettings.GameRatingLevel.values(),
-                selected = gameLimit,
-                labelProvider = { it.displayName },
-                onSelected = onGameLimitChange
-            )
-
-            RatingLimitSelector<ParentalControlsSettings.BookRatingLevel>(
-                icon = Icons.AutoMirrored.Filled.MenuBook,
-                title = "Books & Stories",
-                subtitle = "Control access to Mature or Explicit fiction.",
-                options = ParentalControlsSettings.BookRatingLevel.values(),
-                selected = bookLimit,
-                labelProvider = { it.displayName },
-                onSelected = onBookLimitChange
-            )
-        }
-    }
-}
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
@@ -599,233 +364,8 @@ private fun <T : Enum<T>> RatingLimitSelector(
     }
 }
 
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun TagBlockingCard(
-    selectedCategories: Set<ParentalControlsSettings.TagBlockCategory>,
-    blockedTags: Set<String>,
-    onToggleCategory: (ParentalControlsSettings.TagBlockCategory) -> Unit,
-    onAddTag: (String) -> Unit,
-    onRemoveTag: (String) -> Unit
-) {
-    var newTag by rememberSaveable { mutableStateOf("") }
-    val sortedBlockedTags = remember(blockedTags) { blockedTags.toList().sorted() }
 
-    Card(modifier = Modifier.fillMaxWidth()) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                Icon(
-                    Icons.Default.Sell,
-                    contentDescription = "Media image",
-                    tint = MaterialTheme.colorScheme.primary
-                )
-                Column {
-                    Text(
-                        text = "Tag Restrictions",
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Medium
-                    )
-                    Text(
-                        text = "Block content by sensitive tag categories or specific keywords.",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
 
-            FlowRow(
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                ParentalControlsSettings.TagBlockCategory.values().forEach { category ->
-                    val isSelected = selectedCategories.contains(category)
-                    FilterChip(
-                        selected = isSelected,
-                        onClick = { onToggleCategory(category) },
-                        label = { Text(category.displayName) },
-                        leadingIcon = {
-                            if (isSelected) {
-                                Icon(
-                                    Icons.Default.VisibilityOff,
-                                    contentDescription = "Media image",
-                                    modifier = Modifier.size(18.dp)
-                                )
-                            }
-                        }
-                    )
-                }
-            }
-
-            HorizontalDivider()
-
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text(
-                    text = "Custom blocked tags",
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.Medium
-                )
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    OutlinedTextField(
-                        value = newTag,
-                        onValueChange = { newTag = it },
-                        label = { Text("Add keyword") },
-                        singleLine = true,
-                        modifier = Modifier.weight(1f),
-                        keyboardOptions = KeyboardOptions.Default.copy(
-                            imeAction = ImeAction.Done
-                        ),
-                        keyboardActions = KeyboardActions(onDone = {
-                            val trimmed = newTag.trim()
-                            if (trimmed.isNotEmpty()) {
-                                onAddTag(trimmed)
-                                newTag = ""
-                            }
-                        })
-                    )
-                    Button(
-                        onClick = {
-                            val trimmed = newTag.trim()
-                            if (trimmed.isNotEmpty()) {
-                                onAddTag(trimmed)
-                                newTag = ""
-                            }
-                        },
-                        enabled = newTag.trim().isNotEmpty()
-                    ) {
-                        Text("Add")
-                    }
-                }
-
-                if (sortedBlockedTags.isEmpty()) {
-                    Text(
-                        text = "No custom tags blocked yet.",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                } else {
-                    FlowRow(
-                        horizontalArrangement = Arrangement.spacedBy(8.dp),
-                        verticalArrangement = Arrangement.spacedBy(8.dp)
-                    ) {
-                        sortedBlockedTags.forEach { tag ->
-                            val display = tag.split(" ").joinToString(" ") {
-                                it.replaceFirstChar { ch ->
-                                    if (ch.isLowerCase()) ch.titlecase(Locale.getDefault()) else ch.toString()
-                                }
-                            }
-                            AssistChip(
-                                onClick = { onRemoveTag(tag) },
-                                leadingIcon = {
-                                    Icon(
-                                        Icons.Default.Clear,
-                                        contentDescription = "Media image",
-                                        modifier = Modifier.size(16.dp)
-                                    )
-                                },
-                                label = { Text(display) }
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun SwitchSettingItem(
-    title: String,
-    description: String,
-    checked: Boolean,
-    onCheckedChange: (Boolean) -> Unit,
-    icon: androidx.compose.ui.graphics.vector.ImageVector,
-    enabled: Boolean = true
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Row(
-            modifier = Modifier.weight(1f),
-            horizontalArrangement = Arrangement.spacedBy(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                icon,
-                contentDescription = "Media image",
-                modifier = Modifier.size(20.dp),
-                tint = if (enabled) MaterialTheme.colorScheme.onSurface else MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f)
-            )
-            Column {
-                Text(
-                    title,
-                    style = MaterialTheme.typography.bodyLarge,
-                    color = if (enabled) MaterialTheme.colorScheme.onSurface else MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f)
-                )
-                Text(
-                    description,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-        Switch(
-            checked = checked,
-            onCheckedChange = onCheckedChange,
-            enabled = enabled
-        )
-    }
-}
-
-@Composable
-private fun InfoCard() {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant
-        )
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Icon(
-                    Icons.Default.Info,
-                    contentDescription = "Media image",
-                    tint = MaterialTheme.colorScheme.tertiary
-                )
-                Text(
-                    "How It Works",
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.Medium
-                )
-            }
-            Text(
-                "• Parental controls protect your family from inappropriate content\n" +
-                        "• Set a 4-digit PIN to lock settings and adult content\n" +
-                        "• Choose to hide or lock mature/explicit rated stories\n" +
-                        "• Applies to all fanfiction sites and downloads\n" +
-                        "• PIN uses secure encryption (SHA-256)",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
 
 enum class PinDialogType {
     SET, CHANGE, CLEAR, VERIFY
