+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/settings/SettingsScreen.kt
@@ -855,303 +855,9 @@ fun SettingsScreen(
         }
 }
 
-@Composable
-private fun ThemePickerDialog(
-    currentTheme: ThemePalette,
-    onDismiss: () -> Unit,
-    onSelect: (ThemePalette) -> Unit
-) {
-    // Show enhanced theme picker with all available themes
-    // Note: Only the 6 original themes can be persisted with current ViewModel
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { 
-            Row(
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text("Select Theme") 
-                Text(
-                    text = "15 Available",
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            }
 
-        },
-        text = {
-            LazyColumn(
-                modifier = Modifier.heightIn(max = 500.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                // Show all CleverFerretTheme options grouped by category
-                CleverFerretTheme.entries.groupBy { it.getConfig().category }.forEach { (category, themes) ->
-                    item {
-                        Text(
-                            text = category.name.replace('_', ' '),
-                            style = MaterialTheme.typography.titleSmall,
-                            color = MaterialTheme.colorScheme.primary,
-                            modifier = Modifier.padding(top = 8.dp, bottom = 4.dp)
-                        )
-                    }
-                    
-                    items(themes.size) { index ->
-                        val theme = themes[index]
-                        val config = theme.getConfig()
-                        val isSelected = theme.name == currentTheme.toCleverFerretTheme().name
-                        
-                        Card(
-                            onClick = { 
-                                // Convert CleverFerretTheme to ThemePalette
-                                onSelect(theme.toThemePalette())
-                            },
-                            colors = CardDefaults.cardColors(
-                                containerColor = if (isSelected)
-                                    MaterialTheme.colorScheme.primaryContainer
-                                else
-                                    MaterialTheme.colorScheme.surfaceVariant
-                            )
-                        ) {
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(12.dp),
-                                horizontalArrangement = Arrangement.SpaceBetween,
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Column(modifier = Modifier.weight(1f)) {
-                                    Text(
-                                        text = config.displayName,
-                                        fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Medium,
-                                        style = MaterialTheme.typography.bodyMedium
-                                    )
-                                    Text(
-                                        text = config.description,
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                    
-                                    // Show special features for Ancient Architect themes
-                                    if (theme.isAncientArchitect()) {
-                                        Row(
-                                            horizontalArrangement = Arrangement.spacedBy(4.dp),
-                                            modifier = Modifier.padding(top = 4.dp)
-                                        ) {
-                                            Icon(
-                                                Icons.Default.Star,
-                                                contentDescription = "Media image",
-                                                modifier = Modifier.size(14.dp),
-                                                tint = MaterialTheme.colorScheme.primary
-                                            )
-                                            Text(
-                                                text = "Special",
-                                                style = MaterialTheme.typography.labelSmall,
-                                                color = MaterialTheme.colorScheme.primary
-                                            )
-                                        }
-                                    }
-                                }
-                                if (isSelected) {
-                                    Icon(
-                                        Icons.Default.Check,
-                                        "Selected",
-                                        tint = MaterialTheme.colorScheme.primary
-                                    )
-                                }
-                            }
-                        }
-                    }
 
-                    
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Close")
-            }
-        }
-    )
-}
 
-@Composable
-private fun BottomBarPreferencesCard(
-    availableItems: List<NavigationItem>,
-    preferences: BottomBarPreferences,
-    onOrderChanged: (List<String>, Set<String>) -> Unit,
-    onReset: () -> Unit
-) {
-    MetallicCard {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 12.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Column(modifier = Modifier.weight(1f)) {
-                    Text(
-                        text = "Bottom Bar Items",
-                        style = MaterialTheme.typography.titleSmall,
-                        fontWeight = FontWeight.Medium
-                    )
-                    Text(
-                        text = "Reorder and hide shortcuts to match how you use the app.",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-                TextButton(
-                    onClick = onReset,
-                    enabled = preferences != BottomBarPreferences.Default
-                ) {
-                    Text("Reset")
-                }
-            }
-
-            if (availableItems.isEmpty()) {
-                Text(
-                    text = "Add a library to unlock bottom bar customization.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.padding(horizontal = 16.dp)
-                )
-            } else {
-                BottomBarPreferencesEditor(
-                    availableItems = availableItems,
-                    preferences = preferences,
-                    onOrderChanged = onOrderChanged
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun BottomBarPreferencesEditor(
-    availableItems: List<NavigationItem>,
-    preferences: BottomBarPreferences,
-    onOrderChanged: (List<String>, Set<String>) -> Unit
-) {
-    val orderedItems = remember(availableItems, preferences) {
-        availableItems.orderedForEditor(preferences)
-    }
-    val editorItems = remember(orderedItems, preferences) {
-        mutableStateListOf<BottomBarEditorItem>().apply {
-            orderedItems.forEach { item ->
-                add(
-                    BottomBarEditorItem(
-                        item = item,
-                        visible = item.preferenceId !in preferences.hidden
-                    )
-                )
-            }
-        }
-    }
-
-    val reorderState = rememberReorderableLazyListState(onMove = { from, to ->
-        editorItems.move(from.index, to.index)
-        persistPreferences(editorItems, onOrderChanged)
-    })
-
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxWidth()
-            .heightIn(max = 320.dp)
-            .reorderable(reorderState),
-        state = reorderState.listState,
-        verticalArrangement = Arrangement.spacedBy(8.dp),
-        userScrollEnabled = editorItems.size > 3
-    ) {
-        items(editorItems, key = { it.item.preferenceId }) { editorItem ->
-            BottomBarEditorRow(
-                editorItem = editorItem,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .detectReorderAfterLongPress(reorderState),
-                onVisibilityToggle = { visible ->
-                    val index = editorItems.indexOfFirst { it.item.preferenceId == editorItem.item.preferenceId }
-                    if (index != -1) {
-                        editorItems[index] = editorItems[index].copy(visible = visible)
-                        persistPreferences(editorItems, onOrderChanged)
-                    }
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun BottomBarEditorRow(
-    editorItem: BottomBarEditorItem,
-    modifier: Modifier = Modifier,
-    onVisibilityToggle: (Boolean) -> Unit
-) {
-    Surface(
-        modifier = modifier,
-        shape = RoundedCornerShape(18.dp),
-        tonalElevation = 0.dp,
-        color = MaterialTheme.colorScheme.surfaceVariant
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(40.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(MaterialTheme.colorScheme.surface),
-                contentAlignment = Alignment.Center
-            ) {
-                if (editorItem.visible) {
-                    editorItem.item.selectedIcon()
-                } else {
-                    editorItem.item.icon()
-                }
-            }
-
-            Column(
-                modifier = Modifier.weight(1f),
-                verticalArrangement = Arrangement.spacedBy(2.dp)
-            ) {
-                Text(
-                    text = editorItem.item.label,
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = if (editorItem.visible) FontWeight.Medium else FontWeight.Normal
-                )
-                if (!editorItem.visible) {
-                    Text(
-                        text = "Hidden",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-
-            Switch(
-                checked = editorItem.visible,
-                onCheckedChange = onVisibilityToggle
-            )
-
-            Icon(
-                imageVector = Icons.Default.DragHandle,
-                contentDescription = "Drag to reorder",
-                tint = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
 
 private data class BottomBarEditorItem(
     val item: NavigationItem,
@@ -1174,74 +880,6 @@ private fun persistPreferences(
     onOrderChanged(order, hidden)
 }
 
-@Composable
-private fun MiniPlayerBackgroundDialog(
-    currentMode: com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode,
-    onSelect: (com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode) -> Unit,
-    onDismiss: () -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Mini Player Background") },
-        text = {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 8.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode.values().forEach { mode ->
-                    val isSelected = mode == currentMode
-                    Surface(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clip(MaterialTheme.shapes.medium)
-                            .clickable { onSelect(mode) },
-                        tonalElevation = if (isSelected) 2.dp else 0.dp,
-                        color = if (isSelected)
-                            MaterialTheme.colorScheme.primaryContainer
-                        else
-                            MaterialTheme.colorScheme.surfaceVariant
-                    ) {
-                        Column(
-                            modifier = Modifier.padding(16.dp),
-                            verticalArrangement = Arrangement.spacedBy(4.dp)
-                        ) {
-                            Row(
-                                verticalAlignment = Alignment.CenterVertically,
-                                horizontalArrangement = Arrangement.spacedBy(8.dp)
-                            ) {
-                                Icon(
-                                    imageVector = when (mode) {
-                                        com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode.THEME -> Icons.Default.ColorLens
-                                        com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode.ALBUM_ART_SLIVER -> Icons.Default.Image
-                                        com.universalmedialibrary.data.settings.MiniPlayerBackgroundMode.VISUALIZER -> Icons.Default.GraphicEq
-                                    },
-                                    contentDescription = "Media image"
-                                )
-                                Text(
-                                    text = mode.displayName,
-                                    style = MaterialTheme.typography.titleSmall,
-                                    fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Medium
-                                )
-                            }
-                            Text(
-                                text = mode.description,
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Close")
-            }
-        }
-    )
-}
 
 @Composable
 private fun GearPositionOption(
